requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
eventlet==0.35.2          # async worker for WebSocket
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
